physical-ai-rag-chatbot/pyproject.toml
[project]
name = "physical-ai-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=43.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=1.0.8",
    "google-generativeai>=0.8.4",
    "qdrant-client>=1.12.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2>=2.9.0",
    "pyjwt>=2.8.0",
    "tiktoken>=0.8.0",
    "jiter>=0.8.0",
    "portalocker>=3.0.0",
    "sniffio>=1.3.0",
    "starlette>=0.38.0",
]
